pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "litestar-keycloak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "PyJWT[crypto]>=2.0",
    "testcontainers>=4.14.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "litestar[standard]>=2.0",
    "python-keycloak>=3.0",
    "ruff",
    "mypy",
    "pytest",
    "pytest-cov",
    "pre-commit",
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    "pytest-asyncio>=1.3.0",
    "pytest-xdist>=3.8.0",
    "pytest-timeout>=2.4.0",
    "pytest-benchmark>=5.2.3",
]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
packages = ["src/litestar_keycloak"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v -m 'not integration' --timeout=30"
pythonpath = ["src"]
markers = ["integration: requires Keycloak container"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
